feat(topbar): new version of application drawer
